requirements.txt
python-pptx==0.6.23
anthropic==0.7.8
python-dotenv==1.0.0
pdf2image==1.16.3
requests==2.31.0
pillow==10.4.0